fix: rename lib/index.js to .jsx
as it contains jsx, it should be .jsx file
